+++ packages/dartpip_solver/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+- Fixed an issue with detection of the home directory.
+
 ## 0.0.1
 
 - Initial version.